requirements.txt
asn1crypto=~0.24.0
astroid=~2.2.5
certifi=~2019.6.16
cffi=~1.12.3
colorama=~0.4.1
cryptography=~2.7
future=~0.17.1
isort=~4.3.21
lazy-object-proxy=~1.4.1
mccabe=~0.6.1
peewee=~3.10.0
promise=~2.2.1
pycparser=~2.19
pylint=~2.3.1
python-telegram-bot=~11.1.0
six=~1.12.0
typed-ast=~1.4.0
wrapt=~1.11.2
